testing, linting and refactoring
set up testing with pytest and implemented 2 basic tests. set up linting for PIP8 standards with flake8, since it focuses on simplicity, and for just the microservice on this project I think its enough. also added a justfile for simplifying commands and documented and refactored the code a bit for clairty